refactor(frontend): split playtest debug runtime
